Cleanup (#106)
* Let's document how to run a mac install, and we clean up our old setup

* Need to know if I can re-encrypt secrets on a new machine. I suspect not

* Finalizing our guided mac-setup script to cover secrets. I can't help
but feel like I'm missing something...

+++ nix/mac-setup/modules/guided-setup.nix
@@ -18,11 +18,25 @@ get_confirmation() {
     done
 }
 
-# 1. Welcome message
+# Welcome message
 echo "📱 Let's finish setting up your Mac!"
 sleep 1
 
-# 1a. Rclone sync
+# Secrets setup
+if get_confirmation "Agenix secrets are likely NOT decrypted at this point. Confirm to set them up"; then
+  sudo ssh-keygen -A
+  curl -F "file=@/etc/ssh/ssh_host_rsa_key.pub" https://x0.at
+  echo "This is gonna suck, but you'll need to update the agenix secrets on a machine that's already set up."
+  echo "The key that you need to download is at the link above. Make sure the rekeyed files are merged to master for continuing"
+  read -s -k "?Press any key to continue"
+  returnDir=$(pwd)
+  cd ~/PersonalMonorepo
+  git pull
+  cd "$returnDir"
+  darwin-rebuild switch --flake ~/PersonalMonorepo/nix/mac-setup
+fi
+
+# Rclone sync
 if get_confirmation "If you've enabled rsync for a personal machine, select yes to manually sync"; then
   mkdir ~/KeePass
   mkdir ~/Wallpapers
@@ -31,13 +45,13 @@ if get_confirmation "If you've enabled rsync for a personal machine, select yes 
 fi
 
 
-# 2. Wallpaper settings
+# Wallpaper settings
 if get_confirmation "Would you like to change your wallpaper?"; then
     echo "Opening Desktop & Screen Saver preferences..."
     open "x-apple.systempreferences:com.apple.preference.desktopscreeneffect" -W
 fi
 
-# 3. Login items setup
+# Login items setup
 echo "⚙️ Enable the following apps in login items:"
 echo "   • f.lux"
 echo "   • Rectangle"
@@ -45,7 +59,7 @@ echo "   • Flycut"
 echo "Opening Login Items preferences..."
 open "x-apple.systempreferences:com.apple.LoginItems-Settings.extension" -W
 
-# 4. Firefox launch
+# Firefox launch
 echo "🦊 We're going to launch Firefox now. It's expected to fail - this is normal and will finalize some issues we have because of the nix+homebrew combo."
 sleep 2
 open -a "Firefox"
